put details as separate files
