+++ src/assets/member_photos/team_photo.jpg
