+++ src/components/Contact.js
@@ -1,33 +1,37 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 
 import Header from './Header'
 import Footer from './Footer'
 
 import '../styles/Contact.css';
-import emailjs from '@emailjs/browser';
 
 const Contact = () => {
-
-	const serviceId = process.env.REACT_APP_SERVICE_ID;
-	const templateId = process.env.REACT_APP_TEMPLATE_ID;
-	const publicKey = process.env.REACT_APP_PUBLIC_KEY;
-
 	useEffect(() => {
 		const aboutContainer = document.querySelector('.contactContainer');
 		aboutContainer.classList.add('fade-in-up');
 	}, []);
 
-	const form = useRef();
+	const [formData, setFormData] = useState({
+		name: '',
+		email: '',
+		message: ''
+	});
+	
+	const handleChange = (e) => {
+		setFormData({ ...formData, [e.target.name]: e.target.value });
+	};
+	
+	const handleSubmit = (e) => {
+		e.preventDefault();
 
-	const sendEmail = (e) => {
-	  e.preventDefault(); // prevents the page from reloading when you hit “Send”
-   
-	  emailjs.sendForm(serviceId, templateId, form.current, publicKey)
-		.then((result) => {
-			alert("Thanks for contacting us! We'll be in touch within 1-2 business days");
-		}, (error) => {
-			alert("Oops, something went wrong! Please email us manually instead");
-		});
+		// Handle form submission here. We probably need like node.js and express
+		//
+		//
+
+
+
+		// Clear the form after submission.
+		setFormData({ name: '', email: '', message: '' });
 	};
 	
 	return (
@@ -35,42 +39,50 @@ const Contact = () => {
 			<Header/>
 
 			<div className='contactContainer'>
-				<h2 className='desc'>Email us directly, or contact us using the form anonymously!</h2>
+				<h2>Email us directly, or contact us using the form anonymously!</h2>
 
-				<div className='contact'>
+				<div className='info'>
 					<div className='contactInfo'>
 
 						<h2>Email</h2>
 						<p>robotics@utmsu.ca</p>
 
 						<h2>Address</h2>
-						<p>HackLab, Deerfield Hall,<br/>North Bldg, Mississauga, ON L5L 3E2</p>
+						<p>???</p>
 					</div>
 
-					<form ref={form} onSubmit={sendEmail} className='contactForm'>
+					<form className='contactForm' onSubmit={handleSubmit}>
 						<label>
-							<span>Name</span>
+							Name:
 							<input
-								type="text"
-								name="from_name"
+							type="text"
+							name="name"
+							value={formData.name}
+							onChange={handleChange}
 							/>
 						</label>
 
 						<label>
-							<span>Email</span>
+							Email:
 							<input
-								type="email"
-								name="from_email"
+							type="email"
+							name="email"
+							value={formData.email}
+							onChange={handleChange}
 							/>
 						</label>
+						<br/>
 
 						<label>
-							<span>Message</span>
+							Message:
 							<textarea
-								name="message"
+							name="message"
+							value={formData.message}
+							onChange={handleChange}
 							/>
 						</label>
-						<button type="submit" value="Send">Submit</button>
+						<br />
+						<button type="submit">Submit</button>
 					</form>
 				</div>
 			</div>
@@ -81,4 +93,3 @@ const Contact = () => {
 }
 
 export default Contact;
-
